Filled player_rank() w/ test code
Since I've never used an OrderedDict before, I added in a little bit dummy code in to the player_rank() function to test how items are accessed in this container datatype. The test code just returns the first the first player in the OrderedDict who has a score of 5. 

Clearly, this does not actually solve the problem statement but I wanted to try accessing elements in an OrderedDict before I got too far along