Implemented ranking by list
This actually did not require any new code in the ranking algorithm, but I wanted to check that players would be ranked by their position in the initial list if two players have same score and same # of games played. Test code at bottom and print statements verify this ranking.