Started implementing
I'm assuming the best approach is to use the split() function for strings in Python. I added a split() and print() statement to the code to test how this split() function works in Python. I primarily wanted to see what happens when the character we are using as the split separator is the first character in the string, which will be the case for a direct pathname.